Implemented fast insertion & retrieval using inbuilt cache, added deldoc function

+++ doc-functions.js
@@ -4,18 +4,22 @@ const chalk = require('chalk');
 const msgpack = require('msgpack-lite');
 const lodash = require('lodash');
 
-//File namespace
-let cache = [];
-
 //Create doc
 function newdoc(docname,callback) {
     let msg;
     if(db !== null) {
         let p = `${process.env.LAZLO_SOURCE}/${db}/${docname}.laz`;
-        fs.writeFileSync(p, '');
-        msg = `Document ${docname} created !`;
-        if(callback)
-            callback(chalk.green.bold(msg))
+        if(fs.existsSync(p)) {
+            msg = `Document already exists !`;
+            if(callback)
+                callback(chalk.red.bold(msg))
+        }
+        else {
+            fs.writeFileSync(p, '');
+            msg = `Document ${docname} created !`;
+            if(callback)
+                callback(chalk.green.bold(msg))
+        }
     }
     else {
         msg = `No database selected !`;
@@ -24,6 +28,31 @@ function newdoc(docname,callback) {
     }
 }
 module.exports.newdoc = newdoc;
+
+//Delete Doc
+function deldoc(docname,callback) {
+    let msg;
+    if(db !== null) {
+        let p = `${process.env.LAZLO_SOURCE}/${db}/${docname}.laz`;
+        if(fs.existsSync(p)) {
+            fs.unlinkSync(p);
+            msg = 'Document deleted !';
+            if(callback)
+                callback(chalk.green.bold(msg))
+        }
+        else {
+            msg = 'Document does not exist !';
+            if(callback)
+                callback(chalk.red.bold(msg))
+        }
+    }
+    else {
+        msg = `No database selected !`;
+        if(callback)
+            callback(chalk.red.bold(msg))
+    }
+}
+module.exports.deldoc = deldoc;
 
 //Insert single
 function inserts(docname,input,callback) {
@@ -38,18 +67,32 @@ function inserts(docname,input,callback) {
                 if(callback)
                     callback((chalk.red.bold(msg)));
             }
-            if(fs.statSync(p).size !== 0) {
-                let buffer = fs.readFileSync(p);
-                cache = msgpack.decode(buffer);
+            if(cache.length !== 0 && current_doc === docname) {
+                //Using preloaded cache data for fast insertion
+                cache.push(object);
+                let buffer = msgpack.encode(cache);
+                fs.writeFile(p,buffer,(err) => {
+                    if(err) throw err;
+                });
+                msg = 'Data inserted !';
+                if(callback)
+                    callback((chalk.green.bold(msg)));
+            } else {
+                //Normal Insertion
+                if(fs.statSync(p).size !== 0) {
+                    let buffer = fs.readFileSync(p);
+                    cache = msgpack.decode(buffer);
+                }
+                current_doc = docname;
+                cache.push(object);
+                let buffer = msgpack.encode(cache);
+                fs.writeFile(p,buffer,(err) => {
+                    if(err) throw err;
+                });
+                msg = 'Data inserted !';
+                if(callback)
+                    callback((chalk.green.bold(msg)));
             }
-            cache.push(object);
-            let buffer = msgpack.encode(cache);
-            fs.writeFile(p,buffer,(err) => {
-                if(err) throw err;
-            });
-            msg = 'Data inserted !';
-            if(callback)
-                callback((chalk.green.bold(msg)));
         }
         else {
             msg = 'No database selected or the document does not exist !';
@@ -72,20 +115,36 @@ function insert(docname,input,callback) {
             if(callback)
                 callback((chalk.red.bold(msg)));
         }
-        if(fs.statSync(p).size !== 0) {
-            let buffer = fs.readFileSync(p);
-            cache = msgpack.decode(buffer);
+        if(cache.length !== 0 && current_doc === docname) {
+            //Using preloaded cache data for fast insertion
+            array.forEach((object) => {
+                cache.push(object);
+            });
+            let buffer = msgpack.encode(cache);
+            fs.writeFile(p,buffer,(err) => {
+                if(err) throw err;
+            });
+            msg = 'Data inserted !';
+            if(callback)
+                callback((chalk.green.bold(msg)));
+        } else {
+            //Normal Insertion
+            if(fs.statSync(p).size !== 0) {
+                let buffer = fs.readFileSync(p);
+                cache = msgpack.decode(buffer);
+            }
+            current_doc = docname;
+            array.forEach((object) => {
+                cache.push(object);
+            });
+            let buffer = msgpack.encode(cache);
+            fs.writeFile(p,buffer,(err) => {
+                if(err) throw err;
+            });
+            msg = 'Data inserted !';
+            if(callback)
+                callback((chalk.green.bold(msg)));
         }
-        array.forEach((object) => {
-            cache.push(object);
-        });
-        let buffer = msgpack.encode(cache);
-        fs.writeFile(p,buffer,(err) => {
-            if(err) throw err;
-        });
-        msg = 'Data inserted !';
-        if(callback)
-            callback((chalk.green.bold(msg)));
     }
     else {
         msg = 'No database selected or the document does not exist !';
@@ -100,32 +159,43 @@ function show(docname,callback) {
     let msg;
     let p = `${process.env.LAZLO_SOURCE}/${db}/${docname}.laz`;
     if(db !== null && fs.existsSync(p)) {
-        if(fs.statSync(p).size !== 0) {
-            fs.readFile(p,(err,data) => {
-               if(err) {
-                   msg = 'Data seems to be corrupted !';
-                   if(callback)
-                       callback((chalk.red.bold(msg)));
-               }
-               else {
-                   cache = msgpack.decode(data);
-                   if(typeof cache === "object") {
-                       cache.forEach((object) => {
-                           if (callback)
-                               callback(object);
-                       })
-                   } else {
-                       msg = 'Data seems to be corrupted !';
-                       if(callback)
-                           callback((chalk.red.bold(msg)));
-                   }
-               }
-            });
-        }
-        else {
-            msg = 'Document is empty !';
-            if(callback)
-                callback((chalk.red.bold(msg)));
+        if(cache.length !== 0 && current_doc === docname) {
+            //Fast retrieval using cache
+            cache.forEach((object) => {
+                if (callback)
+                    callback(object);
+            })
+        } else {
+            //Normal retrieval
+            if(fs.statSync(p).size !== 0) {
+                fs.readFile(p,(err,data) => {
+                    if(err) {
+                        msg = 'Data seems to be corrupted !';
+                        if(callback)
+                            callback((chalk.red.bold(msg)));
+                    }
+                    else {
+                        cache = msgpack.decode(data);
+                        if(typeof cache === "object") {
+                            current_doc = docname;
+                            cache.forEach((object) => {
+                                if (callback)
+                                    callback(object);
+                            })
+                        } else {
+                            current_doc = null;
+                            msg = 'Data seems to be corrupted !';
+                            if(callback)
+                                callback((chalk.red.bold(msg)));
+                        }
+                    }
+                });
+            }
+            else {
+                msg = 'Document is empty !';
+                if(callback)
+                    callback((chalk.red.bold(msg)));
+            }
         }
     }
     else {
